Added Purchase Add Method

+++ app/src/main/java/com/example/financehelper/data/model/Purchase.kt
@@ -2,5 +2,5 @@ package com.example.financehelper.data.model
 
 data class Purchase(
     val purchaseName: String,
-    val purchaseCost: Int
+    val purchaseCost: Double
 )
